refactor: move logic to services and repositories, update import paths, remove unnecessary files
Moved logic into services and repositories as requested by QA. Updated import paths and removed unnecessary files.
